+++ example/components/app.post.css
@@ -1,0 +1,11 @@
+.app-container {
+  width: 100%;
+  height: 100%;
+  border: 1px solid red;
+}
+
+.container {
+  padding: 10px;
+  border: 1px solid #61DAFB;
+  margin: 10px;
+}
